refactor(python): promote numpy sentinels to public lib/constants/
Move the sentinel constants (INT32_SENTINEL, FLOAT_SENTINEL,
STRING_SENTINEL, BYTES_SENTINEL, BOOL_SENTINEL) and their dispatch
helper sentinel_for_numpy_dtype out of the private
_internal/utils/numpy_utils/sentinel.py into a new public top-level
package michelangelo.lib.constants.sentinel.

These sentinels are cross-cutting and not owned by any single feature,
so they belong in a lib-wide public constants package (mirroring
model_manager/constants/ and the internal shared/constants/sentinel.py).
lib/constants/__init__.py re-exports the public API.

numpy_utils.pad and the trainer/torch collate helpers/tests now import
from the public location; the numpy_utils package no longer re-exports
the sentinels. The sentinel test moves to lib/constants/tests/.
