+++ python/michelangelo/lib/_internal/utils/numpy_utils/__init__.py
@@ -6,26 +6,12 @@ from michelangelo.lib._internal.utils.numpy_utils.pyarrow_conversion import (
     numpy_to_pyarrow,
     pyarrow_to_numpy,
 )
-from michelangelo.lib._internal.utils.numpy_utils.sentinel import (
-    BOOL_SENTINEL,
-    BYTES_SENTINEL,
-    FLOAT_SENTINEL,
-    INT32_SENTINEL,
-    STRING_SENTINEL,
-    sentinel_for_numpy_dtype,
-)
 from michelangelo.lib._internal.utils.numpy_utils.type import infer_dtype
 
 __all__ = [
-    "BOOL_SENTINEL",
-    "BYTES_SENTINEL",
-    "FLOAT_SENTINEL",
-    "INT32_SENTINEL",
-    "STRING_SENTINEL",
     "assemble_output_table",
     "infer_dtype",
     "numpy_to_pyarrow",
     "pad_ragged_tensor",
     "pyarrow_to_numpy",
-    "sentinel_for_numpy_dtype",
 ]
