+++ python/michelangelo/lib/constants/sentinel.py
