+++ python/michelangelo/lib/trainer/torch/tests/test_data_collate_functions.py
@@ -13,7 +13,7 @@ import numpy as np
 import pytest
 import torch
 
-from michelangelo.lib._internal.utils.numpy_utils import (
+from michelangelo.lib.constants.sentinel import (
     FLOAT_SENTINEL,
     INT32_SENTINEL,
 )
